chore: add shared view helpers module
